ML-6127: Workaround for broken application state tracking on legacy iOS.
When the Xcode 9/iOS 11 version of XCTest is used and the target device
is a previous version of iOS, XCTest falls back to a legacy code path
for -[XCUIApplication applicationWithPID:] due to the non-existence of
the _XCT_requestBundleIDForPID:reply: RPC on that platform. Because it
creates an XCUIApplication with a dummy bundle ID and registers it for
application state tracking, the created application object will never be
considered to be "running" and element queries will fail. This breaks
scenarios involving navigating out of the starting app to some other
app and interacting with it.

The workaround overrides the values returned by XCUIApplication that are
needed by XCUIElementQuery to service requests to it, such that the
application will be considered running. This isn't technically correct,
but without a way to get the bundle ID of an arbitrary process, there's
no actual correct way to do the state tracking, and this is better than
it completely not working.