ML-6080: Handle AX error -25205 in attribute values.
This string is returned by testmanagerd when accessiblity attributes
which are not supported by the target accessibility element are
requested. It appears there may be some kind of problem in
XCElementSnapshot's "fault in" handler when targeting legacy iOS
versions from iOS 11 XCTest that's causing it to do this. The root cause
isn't fully confirmed yet, and there's what appears to be a ton of
spurious roundtrips to testmanagerd happening that need to be resolved
as well, but this at least fixes some of the more glaring behavior.